+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [3.1.1] / 2025-02-19
-### Fixed
-- RollBack - Ignore create/delete error in `BucketKey`.
+### Updated
+- Add `GetBucketKey` to create unique bucket key with hash.
 
 ## [3.1.0] / 2025-01-22
 ### Features